PP-9955 refactor `config` structure
- refactor `config` structure to follow GDS' project directory structure,
  move config stuff from root dir to be under `config`
