+++ app/payment/pre-payment.controller.it.test.js
@@ -5,7 +5,7 @@ const config = require('../../config')
 const cheerio = require('cheerio')
 const nock = require('nock')
 const supertest = require('supertest')
-const { getApp } = require('../../server')
+const { getApp } = require('../../config/server')
 const { createAppWithSession } = require('../../test/test-helpers/mock-session')
 const productFixtures = require('../../test/fixtures/product.fixtures')
 const serviceFixtures = require('../../test/fixtures/service.fixtures')
